+++ src/apps/front/src/features/app/App.scss
@@ -2,6 +2,5 @@
 @import './styles/reset';
 @import './styles/normalize';
 @import './styles/system/color';
-@import '../feed/feed';
 @import '../auth/join/components/join';
 @import '../auth/login/components/login';
