--- src/apps/front/src/features/chat/Chat.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Chat = () => {
-  return <div>chat</div>;
-};
-
-export default Chat;
