--- src/apps/front/src/features/feed/Explore.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const explore = () => {
-  return <div>explore</div>;
-};
-
-export default explore;
