--- src/apps/front/src/features/feed/Feed.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import Gnb from './components/Gnb';
-import Lnb from './components/Lnb';
-
-const Feed = () => {
-  return (
-    <div>
-      <Lnb></Lnb>
-      <Gnb></Gnb>
-    </div>
-  );
-};
-
-export default Feed;
