--- src/apps/front/src/features/feed/components/Alertpop.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Alertpop = () => {
-  return <div>Alertpop</div>;
-};
-
-export default Alertpop;
