--- src/apps/front/src/features/feed/components/Post.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Post = () => {
-  return <div>Post</div>;
-};
-
-export default Post;
