--- src/apps/front/src/features/imagination/Imagination.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Imagination = () => {
-  return <div>Imagination</div>;
-};
-
-export default Imagination;
