--- src/apps/front/src/features/leader/Leader.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Leader = () => {
-  return <div>leader</div>;
-};
-
-export default Leader;
